server/requirements.txt
fastapi[standard]
cloudinary
python-dotenv
uvicorn